requirements.txt
pyeapi==0.8.1
PyYAML
pytest
requests
click
rich
jsonschema